Model acts as base abstract class & KerasModel as child abstract class.
