+++ .gitignore
@@ -4,5 +4,5 @@ logs
 __pycache__
 issues.txt
 t.py
-model/saved_models
+model/keras_model/saved_models
 data/scaler.gz
