--- model/constants.py
@@ -1,4 +1,0 @@
-SEQ_LEN = 60
-BATCH_SIZE = 64
-STEP = 1
-SAVED_MODELS_BASE_PATH = 'saved_models'
