+++ utils/utils.py
@@ -1,5 +1,4 @@
 import datetime as dt
-from model.constants import SEQ_LEN
 import os
 
 import pandas as pd
@@ -17,9 +16,9 @@ class InvalidTickerError(Exception):
 
 
 class InvalidPredictionDateError(Exception):
-    def __init__(self, pred_date: dt.date, df_date: dt.date, lower: bool = False) -> None:
+    def __init__(self, pred_date: dt.date, df_date: dt.date, seq_len: int, lower: bool = False) -> None:
         if lower:
-            msg = f"We have data till {df_date} and model needs sequences of length {SEQ_LEN} to predict. So model can predict only for dates starting from {df_date + dt.timedelta(days=SEQ_LEN)}. But you called model's predict function with this prediction date: {pred_date}"
+            msg = f"We have data till {df_date} and model needs sequences of length {seq_len} to predict. So model can predict only for dates starting from {df_date + dt.timedelta(days=seq_len)}. But you called model's predict function with this prediction date: {pred_date}"
         else:
             msg = f"We have data till {df_date}, so model can predict only till {df_date + dt.timedelta(days=1)}. But you called model's predict function with this prediction date: {pred_date}"
         super().__init__(msg)
@@ -60,7 +59,7 @@ def get_date_from_string(date: str) -> dt.date:
         raise InvalidDateError(date)
 
 
-def get_prediction_date(df: pd.DataFrame, date: str = None):
+def get_prediction_date(df: pd.DataFrame, seq_len: int, date: str = None):
     if date is None:
         pred_date = dt.datetime.now().date()
         # timezone = pytz.timezone("Asia/Kolkata")
@@ -69,17 +68,17 @@ def get_prediction_date(df: pd.DataFrame, date: str = None):
 
     df_first_date = df.index[0].date()
     df_last_date = df.index[-1].date()
-    return get_correct_pred_date(pred_date, df_first_date, df_last_date)
+    return get_correct_pred_date(pred_date, df_first_date, df_last_date, seq_len)
 
 
-def get_correct_pred_date(pred_date: dt.date, df_first_date: dt.date, df_last_date: dt.date):
+def get_correct_pred_date(pred_date: dt.date, df_first_date: dt.date, df_last_date: dt.date,  seq_len: int):
     """
     If df_last_date is Fri and date is Sat or Sun then the actual pred_date is of next Mon
 
     If df_last_date >= date - dt.timedelta(days=1) then we surely have the data for that pred_date
     """
 
-    if df_first_date + dt.timedelta(days=SEQ_LEN) > pred_date:
+    if df_first_date + dt.timedelta(days=seq_len) > pred_date:
         raise InvalidPredictionDateError(pred_date, df_first_date, lower=True)
 
     # If df_last_date is Fri and pred_date is greater than next Mon
